requirements.txt
Flask==3.0.0
gunicorn==21.2.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-CORS==4.0.0
python-chess==1.999
psycopg2-binary==2.9.9
werkzeug==3.0.1
python-dotenv==1.0.1
requests==2.31.0
